+++ src/test/java/org/morphix/reflection/ConstructorsNewInstanceTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test;
  *
  * @author Radu Sebastian LAZIN
  */
-class ConstructorsTest {
+class ConstructorsNewInstanceTest {
 
 	public static class A {
 		// empty
@@ -97,5 +97,4 @@ class ConstructorsTest {
 				+ " the class object represents an abstract class, an interface,"
 				+ " an array class, a primitive type, or void: "));
 	}
-
 }