+++ src/test/java/org/morphix/reflection/ConstructorsSafeGetDeclaredConstructorWithParamsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -15,20 +15,21 @@ package org.morphix.reflection;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 import java.lang.reflect.Constructor;
 
 import org.junit.jupiter.api.Test;
 
 /**
- * Test class for {@link Constructors#newInstance(Class)}.
+ * Test class for {@link Constructors.Safe#getDeclared(Class)}.
  *
  * @author Radu Sebastian LAZIN
  */
-class ConstructorsGetDeclaredConstructorWithParamsTest {
+class ConstructorsSafeGetDeclaredConstructorWithParamsTest {
 
 	public static class B {
+
 		private final int i;
 		private String s;
 
@@ -53,7 +54,7 @@ class ConstructorsGetDeclaredConstructorWithParamsTest {
 	public static class C {
 		private final long x;
 
-		private C(final long x) {
+		private C(final long x) { // NOSONAR used by reflection
 			this.x = x;
 		}
 
@@ -64,14 +65,14 @@ class ConstructorsGetDeclaredConstructorWithParamsTest {
 
 	@Test
 	void shouldFindConstructor() {
-		Constructor<B> constructor = Constructors.getDeclared(B.class, int.class, String.class);
+		Constructor<B> constructor = Constructors.Safe.getDeclared(B.class, int.class, String.class);
 
 		assertNotNull(constructor);
 	}
 
 	@Test
 	void shouldCreateNewInstanceWithConstructor() {
-		Constructor<B> constructor = Constructors.getDeclared(B.class, int.class, String.class);
+		Constructor<B> constructor = Constructors.Safe.getDeclared(B.class, int.class, String.class);
 		B a = Constructors.IgnoreAccess.newInstance(constructor, 10, "test");
 
 		assertNotNull(a);
@@ -81,20 +82,21 @@ class ConstructorsGetDeclaredConstructorWithParamsTest {
 
 	@Test
 	void shouldThrowExceptionIfNoConstructorIsFoundWithGivenParameters() {
-		ReflectionException e = assertThrows(ReflectionException.class, () -> Constructors.getDeclared(B.class, String.class));
-		assertThat(e.getMessage(),
-				equalTo("No constructor found for class: " + B.class.getCanonicalName() + " with parameters: [class java.lang.String]"));
+		Constructor<B> constructor = Constructors.Safe.getDeclared(B.class, String.class);
+
+		assertNull(constructor);
 	}
 
 	@Test
 	void shouldThrowExceptionIfNoConstructorIsFoundWithNullParameters() {
-		ReflectionException e = assertThrows(ReflectionException.class, () -> Constructors.getDeclared(B.class, (Class<?>[]) null));
-		assertThat(e.getMessage(), equalTo("No constructor found for class: " + B.class.getCanonicalName() + " with parameters: none"));
+		Constructor<B> constructor = Constructors.Safe.getDeclared(B.class, (Class<?>[]) null);
+
+		assertNull(constructor);
 	}
 
 	@Test
 	void shouldReturnPrivateConstructor() {
-		Constructor<C> constructor = Constructors.getDeclared(C.class, long.class);
+		Constructor<C> constructor = Constructors.Safe.getDeclared(C.class, long.class);
 
 		assertNotNull(constructor);
 	}